# AssessFlow_System_Design_Pack_v1.0/07_AssessFlow_OpenAPI_v1.0.yaml
openapi: 3.1.0
info:
  title: AssessFlow API
  version: 1.0.0
  description: Core REST contract for the AssessFlow employee assessment module. Numeric scoring is optional and disabled
    by default in the MVP.
servers:
- url: /api/v1/assessflow
security:
- bearerAuth: []
...
paths:
  /cases:
    get:
      tags:
      - Cases
      summary: List assessment cases
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - in: query
        name: status
        schema:
          type: string
      - in: query
        name: assessmentReason
        schema:
          $ref: '#/components/schemas/AssessmentReason'
      responses:
        '200':
          description: Paged case list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AssessmentCase'
                  total:
                    type: integer
    post:
      tags:
      - Cases
      summary: Create a draft assessment case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentCase'
        '422':
          $ref: '#/components/responses/BusinessRuleFailed'
  /cases/{caseId}:
    parameters:
    - $ref: '#/components/parameters/CaseId'
    get:
      tags:
      - Cases
      summary: Get assessment case detail
      responses:
        '200':
          description: Case detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentCase'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - Cases
      summary: Edit a draft or changes-requested case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetPositionId:
                  type: string
                  format: uuid
                targetLevelId:
                  type:
                  - string
                  - 'null'
                  format: uuid
                justification:
                  type: string
                priority:
                  type:
                  - string
                  - 'null'
                requestedCompletionDate:
                  type:
                  - string
                  - 'null'
                  format: date
                expectedVersion:
                  type: integer
      responses:
        '200':
          description: Updated
        '409':
          $ref: '#/components/responses/Conflict'
  /cases/{caseId}/submit:
    post:
      tags:
      - Cases
      summary: Submit case for eligibility review
      parameters:
      - $ref: '#/components/parameters/CaseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionCommand'
      responses:
        '200':
          description: Submitted
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleFailed'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CaseId:
      in: path
      name: caseId
      required: true
      schema:
        type: string
        format: uuid
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
  schemas:
    AssessmentReason:
      type: string
      enum:
      - INTERNAL_MOBILITY
      - PROMOTION
      - ROLE_REALIGNMENT
...
    AssessmentCase:
      type: object
      required:
      - id
      - caseCode
      - employeeId
      - assessmentReason
      - targetPositionId
      - stage
      - status
      - scoringEnabled
      - version
      properties:
        id:
          type: string
          format: uuid
        caseCode:
          type: string
        employeeId:
          type: string
          format: uuid
        requesterUserId:
          type: string
          format: uuid
        assessmentReason:
          $ref: '#/components/schemas/AssessmentReason'
        currentPositionId:
          type:
          - string
          - 'null'
          format: uuid
        currentLevelId:
          type:
          - string
          - 'null'
          format: uuid
        targetPositionId:
          type: string
          format: uuid
        targetLevelId:
          type:
          - string
          - 'null'
          format: uuid
        complexityBand:
          type:
          - string
          - 'null'
        justification:
          type: string
        stage:
          type: string
        status:
          type: string
        scoringEnabled:
          type: boolean
          default: false
        version:
          type: integer
    CreateCaseRequest:
      type: object
      required:
      - employeeId
      - assessmentReason
      - targetPositionId
      - justification
      properties:
        employeeId:
          type: string
          format: uuid
        assessmentReason:
          $ref: '#/components/schemas/AssessmentReason'
        targetPositionId:
          type: string
          format: uuid
        targetLevelId:
          type:
          - string
          - 'null'
          format: uuid
        justification:
          type: string
          minLength: 1
        priority:
          type:
          - string
          - 'null'
        requestedCompletionDate:
          type:
          - string
          - 'null'
          format: date
    VersionCommand:
      type: object
      required:
      - expectedVersion
      properties:
        expectedVersion:
          type: integer
          minimum: 1
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            correlationId:
              type:
              - string
              - 'null'
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Invalid transition or version conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BusinessRuleFailed:
      description: Business rule failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'